rename 'service.rs' to 'block_producer.rs'
