+++ crates/block-producer/src/lib.rs
@@ -1,13 +1,13 @@
+mod block_producer;
 mod context;
 mod error;
-mod service;
 
 pub mod rpc;
 pub mod services;
 pub mod types;
 
+pub use block_producer::{BlockProducer, run};
 pub use context::BlockProducerContext;
-pub use service::{BlockProducer, run};
 
 pub mod prelude {
     pub use crate::{
